Refactor api reducers into shared hook
