+++ apps/core-frontend/src/hooks/useMe.hook.ts
@@ -1,36 +1,6 @@
-import * as React from 'react';
-
 import { UserClient } from '@nrcno/core-clients';
 import { User } from '@nrcno/core-models';
-
-enum SubmitStatus {
-  IDLE = 'idle',
-  SUBMITTING = 'submitting',
-  SUCCESS = 'success',
-  ERROR = 'error',
-}
-
-type State = {
-  status: SubmitStatus;
-  data?: User;
-  error?: Error;
-};
-
-type Action =
-  | {
-      type: SubmitStatus.IDLE;
-    }
-  | {
-      type: SubmitStatus.SUBMITTING;
-    }
-  | {
-      type: SubmitStatus.SUCCESS;
-      data: User;
-    }
-  | {
-      type: SubmitStatus.ERROR;
-      data: Error;
-    };
+import { SubmitStatus, useApiReducer } from '@nrcno/core-prm-components';
 
 export type MeState = {
   getMe: () => Promise<any>;
@@ -45,43 +15,17 @@ export const defaultMeState: MeState = {
 };
 
 export const useMe = (client: UserClient): MeState => {
-  const [state, dispatch] = React.useReducer(
-    (state: State, action: Action) => {
-      switch (action.type) {
-        case SubmitStatus.SUBMITTING:
-          return { ...state, status: SubmitStatus.SUBMITTING };
-        case SubmitStatus.SUCCESS:
-          return {
-            ...state,
-            status: SubmitStatus.SUCCESS,
-            data: action.data,
-          };
-        case SubmitStatus.ERROR:
-          return {
-            ...state,
-            status: SubmitStatus.ERROR,
-            error: action.data,
-          };
-        case SubmitStatus.IDLE:
-          return { ...state, status: SubmitStatus.IDLE };
-        default:
-          return state;
-      }
-    },
-    {
-      status: SubmitStatus.IDLE,
-    },
-  );
+  const [state, actions] = useApiReducer<User>();
 
   const getMe = async () => {
     try {
-      dispatch({ type: SubmitStatus.SUBMITTING });
+      actions.submitting();
       const me = await client.getMe();
-      dispatch({ type: SubmitStatus.SUCCESS, data: me });
+      actions.success(me);
     } catch (error) {
       const err =
         error instanceof Error ? error : new Error('An error occurred');
-      dispatch({ type: SubmitStatus.ERROR, data: err });
+      actions.error(err);
     }
   };
