+++ libs/prm-components/src/index.ts
@@ -2,4 +2,4 @@ export * from './lib/pages';
 // export * from './lib/components';
 export * from './lib/prm.context';
 export * from './lib/config';
-export * from './lib/types';
+export * from './lib/hooks';
