+++ libs/prm-components/src/lib/hooks/index.ts
@@ -1,1 +1,1 @@
-export * from './useCreateEntity.hook';
+export * from './useApiReducer.hook';
