+++ libs/prm-components/src/lib/prm.context.tsx
@@ -1,14 +1,13 @@
 import * as React from 'react';
 import { AxiosInstance } from 'axios';
 import { PrmClient } from '@nrcno/core-clients';
-import { useLoaderData, useParams } from 'react-router-dom';
 import { EntityType, EntityTypeSchema } from '@nrcno/core-models';
 
 import {
   CreateEntityState,
   defaultCreateEntityState,
   useCreateEntity,
-} from './hooks';
+} from './hooks/useCreateEntity.hook';
 import {
   ReadEntityState,
   defaultReadEntityState,
