--- libs/prm-components/src/lib/types/index.ts
@@ -1,1 +1,0 @@
-export * from './reducer.types';
